add: base file field
